refactor(studio): decouple anonymizer role model defaults from the tab
Extract the workspace model list and the per-role defaulting effect into
useAnonymizerModels/useDefaultRoleModels, and run the defaulting from a
RoleModelDefaults child of the form provider. Previously the defaults were
seeded by an effect inside ModelSettingsSection, so they only applied
because that tab stays mounted; ModelSettingsSection is now pure render.

Groundwork for wrapping the tab panels in <Activity>, which is not safe
here yet: react-hook-form's useWatch subscribes in an effect, and Activity
tears down effects while hidden, so the inactive tab renders stale values.
